Merged PR 1654396: [Update procedure] Support Unsharded Remote collections
### Does this PR have any customer impact?
Yes

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
Feature

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
No

### Are you introducing any new config? If yes, do you have tests with and without them being set?
No

### ChangeLog (Refer [Template](../oss/CHANGELOG.md))

### Description

For the Update stored procedure for an unsharded collection on a remote node, we would typically delegate the entire request to the worker and process it there.

With a SPROC, we can't delegate the whole thing to the worker. At the same time we can't make it request by request doing RPC to the worker. This change strikes a balance between the two, where we get batches of batchWriteInternalSubTransactionCount (~512) requests, and call update worker with it. Each batch is run as a single transaction on the worker, parsed, evaluated and Commited within the function. And then the outer Sproc commits the whole thing.

This balances between minimizing RPCs and Committing Frequently.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces support for unsharded remote collections in the update procedure, enhancing the functionality of batch updates. Key changes include:
- `update.c`: Added functions to deserialize writer errors and upserted results, serialize update batch parameters, and process batch updates non-transactionally for unsharded collections.
- `commands_update_bulk.out`: Updated expected test outputs to reflect changes in batch update behavior.
- `commands_update_bulk.sql`: Modified SQL tests to disable local writes and test batch update functionality with remote collections.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #4093126