Merged PR 1689621: Change default for localhost to false
### Does this PR have any customer impact?

No

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)

Bugfix

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)

No

### Are you introducing any new config? If yes, do you have tests with and without them being set?

No

### ChangeLog (Refer [Template](../oss/CHANGELOG.md))

### Description

Change default for localhost to false

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a configuration change that updates the default behavior for local host usage.

#### PR Summary
The changes modify the configuration setup to return false by default for using localhost, ensuring that the application does not use the local host unless explicitly configured.
- `oss/pg_documentdb_gw/src/configuration/setup.rs`: Changed the `use_local_host` method to use `unwrap_or(false)` instead of `unwrap_or(true)`.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->