Merged PR 1722022: [OSS] Introduce documentdb_rum under the internal folder
### Does this PR have any customer impact?
No

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
Feature

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
No

### Are you introducing any new config? If yes, do you have tests with and without them being set?
No
### ChangeLog (Refer [Template](../oss/CHANGELOG.md))

### Description

Introduce documentdb_rum which is a fork of the OSS Rum index that will be used for advanced documentdb scenarios.

This change:
1) introduces the access method and forks the code (as-is) makes it build in the repo.
2) Removes any tsvector/tsquery, arrayutils/rum btree  logic that was there before
3) Deletes custom logic that was built for PG9.6 -> pg14
4) Trims the SQL file to only have hte Access method.

This change does not wire up the RUM index with documentdb or provide the modifications needed for the documentdb scenarios (yet). This will be done in subsequent changes.