Merged PR 1724016: [Composite] Allow side by side non composite and composite index with composite as default
### Does this PR have any customer impact?

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)

### Are you introducing any new config? If yes, do you have tests with and without them being set?

### ChangeLog (Refer [Template](../oss/CHANGELOG.md))

### Description

[Composite] Allow side by side non composite and composite index with composite opclass as default

We treat indexes with the `enableCompositeTerm` flag set to true as non equivalent to indexes with this flag set to false. That means that we can have these indexes:

{a: 1}
{a: 1}, enableCompositeTerm: true

However, when the GUC `defaultUseCompositeOpClass` is set to on, we don't track the `enableCompositeTerm` flag as part of the index spec if the user doesn't provide it in the spec, so we deem the indexes are equivalent and not allowing to have a composite index one the same  field.

This is helpful for scenarios where we wanna migrate to using a composite index but we have large data, so the index build will probably take some time, but we still wanna have an index for ongoing workloads.

----
#### AI description  (iteration 1)
#### PR Classification
This PR is an API change that enhances index creation functionality by allowing non composite and composite indexes to coexist with composite as the default.

#### PR Summary
The PR refines the index creation logic to support side-by-side non composite and composite indexes by applying new configuration flags and error-handling mechanisms, and it updates corresponding tests to verify the behavior.
- `src/commands/create_indexes.c`: Modified composite index validation by introducing conditional error reporting and setting a new flag based on configuration.
- `include/commands/create_indexes.h`: Added the `canSupportCompoundTerm` field to indicate key support for composite terms.
- `src/test/regress/expected/bson_composite_index_tests_insert_terms.out` & `src/test/regress/sql/bson_composite_index_tests_insert_terms.sql`: Updated test scripts to validate the creation of both non composite and composite indexes with the same key.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->