Merged PR 1732387: [Bug-Fix] Fixing Type of data we send as Response of udpate
### Does this PR have any customer impact?

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)

### Are you introducing any new config? If yes, do you have tests with and without them being set?

### ChangeLog (Refer [Template](../oss/CHANGELOG.md))

### Description
Previously, we were always sending int64 for rowMatched and rowModified, but some drivers expect int32. Now, the response adjusts based on the value: if the data fits within int32, it sends int32; otherwise, it uses int64.

some drivers expect int32 which lead to incorrect result, below are two runs the later one is with my fix
![image.png](https://msdata.visualstudio.com/ba574a88-a171-48e0-8fcb-5fef6d23739c/_apis/git/repositories/501a5e7e-ffdc-4d1e-a5e6-910f73c6eb9a/pullRequests/1732387/attachments/image.png)