Merged PR 1774203: Fix up crash in the entryFindItem path
### Does this PR have any customer impact?
Yes

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
Bugfix

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
No

### Are you introducing any new config? If yes, do you have tests with and without them being set?
No

### ChangeLog (Refer [Template](../oss/CHANGELOG.md))

### Description

Fix up crash in the entryFindItem path - it would be the same fix as what was observed in the entryGetNextItem path.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix: This PR resolves a crash in the entryFindItem path caused by page splits.

#### PR Summary
This change adds logic in the entryFindItem function to repeatedly locate the leftmost page when a page split occurs, ensuring proper buffer locking and preventing crashes.
- `oss/internal/pg_documentdb_extended_rum/src/rumget.c`: Introduced a while loop to refind the leftmost page by checking for non-leaf pages and re-locking the new buffer during entryFindItem.
- `pgmongo_rum/src/rumget.c`: Made similar modifications by adding the `Page page;` declaration and corresponding loop to handle page splits effectively.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->