Merged PR 1806558: Correctly generating the activity_id including the original requestId
Embed the requestId in the activity_id making it easier to trace a complete request from the moment the connection was created all the way up until it is closed.

Also, we were not including the activity_id in all the places that we could making tracing difficult. As well, we performed some log cleanup to only include useful information in the messages.

```
2025-09-17T20:28:34.582Z New TCP connection established with activity id: '77a5e6b3-17f2-4102-9f73-8cd16786a9db'
2025-09-17T20:28:36.482Z Processing connection: '77a5e6b3-17f2-4102-9f73-8cd100000013'. Client IP: '[::ffff:127.0.0.1]:60126', HostName: 'localhost'
2025-09-17T20:28:36.483Z Connection with id '77a5e6b3-17f2-4102-9f73-8cd100000013' closed.
```

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a bug fix that improves logging and connection tracking by correctly embedding the original requestId in the activity_id.

Related work items: #4584662

Signed-off-by: Shuai Tian <shuaitian@microsoft.com>